requirements/production.txt
asgiref==3.6.0
Django==4.1.5
django-auth-ldap==4.1.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
python-ldap==3.4.3
pytz==2022.7
mysqlclient==2.1.1
sqlparse==0.4.3
tzdata==2022.7